requirements.txt

faster-whisper
google-generativeai
pydantic
python-dotenv
